# openapi.yaml
openapi: 3.0.3
info:
  title: RESIZR - Image Processing API
  description: |
    **RESIZR** is a high-performance image processing service that provides RESTful APIs for image upload, processing, and delivery. 
    
    ## Features
    - Multi-resolution image processing with configurable resize algorithms (smart-fit, crop, stretch)
    - Support for JPEG, PNG, GIF, and WebP formats
    - Redis-based metadata storage and caching
    - S3-compatible storage backend
    - Rate limiting and security features
    
    ## Authentication
    RESIZR supports optional API key-based authentication with two permission levels:
    
    - **Read-Write Keys**: Can upload images and access all read operations
    - **Read-Only Keys**: Can only access download and info operations
    
    Authentication is disabled by default and can be enabled via configuration.
    When enabled, API keys must be provided in the `X-API-Key` header (configurable).
    
    ## Rate Limits
    - Upload: 10 requests/minute per IP
    - Download: 100 requests/minute per IP
    - Info: 50 requests/minute per IP
    
    ## CORS Support
    Cross-Origin Resource Sharing (CORS) is configurable:
    - Production: Restricted to configured allowed origins
    - Development: Allows all origins for testing
    - Configurable through environment variables
    
    ## Error Handling
    All errors follow a consistent JSON format with appropriate HTTP status codes.
    
  version: 0.0.1
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/v1/images/{id}/{resolution}/presigned-url:
    get:
      tags:
        - Images
      summary: Generate presigned URL
      description: |
        Generate a temporary presigned URL for direct access to image files stored in S3.
        
        **Features:**
        - Direct client-side access to S3 without server proxy
        - Configurable expiration time (default: 1 hour, max: 7 days)
        - Support for all available image resolutions
        - Reduces server bandwidth and improves performance
        - Consistent with existing download endpoint structure
        
        **Use Cases:**
        - Mobile app integration with direct S3 access
        - Temporary image sharing with expiration
        - Client-side downloads for large images
        - CDN integration and caching optimization
        
      operationId: generatePresignedURL
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - $ref: '#/components/parameters/Resolution'
        - name: expires_in
          in: query
          required: false
          description: "URL expiration time in seconds (default: 3600, max: 604800)"
          schema:
            type: integer
            minimum: 1
            maximum: 604800
            default: 3600
          example: 3600
      responses:
        '200':
          description: Presigned URL generated successfully
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Unique request identifier for tracing
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: CORS allowed origin
            Access-Control-Expose-Headers:
              schema:
                type: string
              description: Headers exposed to the client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedURLResponse'
              example:
                url: "https://bucket.s3.amazonaws.com/images/f47ac10b-58cc-4372-a567-0e02b2c3d479/thumbnail.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=..."
                expires_at: "2025-09-12T16:30:00Z"
                expires_in: 3600
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/images/{id}/original:
    get:
      tags:
        - Images
      summary: Download original image
      description: |
        Download the original uploaded image file.
        
        **Features:**
        - Streaming download for large files
        - Proper caching headers for CDN compatibility
        - ETag support for conditional requests
        - Range request support for partial downloads
        
      operationId: downloadOriginal
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/Range'
      responses:
        '200':
          description: Original image file
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: CORS allowed origin
            Access-Control-Expose-Headers:
              schema:
                type: string
              description: Headers exposed to the client
            Content-Type:
              schema:
                type: string
              description: MIME type of the image
              example: "image/jpeg"
            Content-Length:
              schema:
                type: integer
              description: Size of the image in bytes
              example: 2048000
            Cache-Control:
              schema:
                type: string
              description: Caching directives
              example: "public, max-age=31536000, immutable"
            ETag:
              schema:
                type: string
              description: Entity tag for caching
              example: '"abc123def456"'
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Unique request identifier for tracing
          content:
            image/*:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (range request)
          headers:
            Content-Type:
              schema:
                type: string
              example: "image/jpeg"
            Content-Range:
              schema:
                type: string
              example: "bytes 0-1023/2048000"
            Content-Length:
              schema:
                type: integer
              example: 1024
          content:
            image/*:
              schema:
                type: string
                format: binary
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/images/{id}/thumbnail:
    get:
      tags:
        - Images
      summary: Download thumbnail
      description: |
        Download a 150x150 pixel thumbnail version of the image.
        
        **Processing:**
        - Smart-fit algorithm maintains aspect ratio
        - White background for letterboxing/pillarboxing
        - Optimized compression for fast loading
        
      operationId: downloadThumbnail
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Thumbnail image (150x150)
          headers:
            Content-Type:
              schema:
                type: string
              example: "image/jpeg"
            Content-Length:
              schema:
                type: integer
              example: 8192
            Cache-Control:
              schema:
                type: string
              example: "public, max-age=31536000, immutable"
            ETag:
              schema:
                type: string
              example: '"thumbnail-abc123"'
          content:
            image/*:
              schema:
                type: string
                format: binary
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/images/{id}/{resolution}:
    get:
      tags:
        - Images
      summary: Download custom resolution
      description: |
        Download a custom resolution version of the image.
        
        **Resolution Format:**
        - Must match pattern: WIDTHxHEIGHT (e.g., "800x600", "1200x900")
        - Both width and height must be positive integers
        - Maximum dimension (hard cap): 8,192 pixels. Default is 4,096 and configurable via IMAGE_MAX_WIDTH and IMAGE_MAX_HEIGHT up to the hard cap.
        - Resolution must have been requested during upload
        
        **Processing:**
        - Smart-fit algorithm maintains aspect ratio
        - White background for letterboxing/pillarboxing
        - Quality optimized for resolution size
        
      operationId: downloadCustomResolution
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - name: resolution
          in: path
          required: true
          description: |
            Custom resolution in WIDTHxHEIGHT format (e.g., "800x600").
            Must be a resolution that was requested during upload.
          schema:
            type: string
            pattern: '^[1-9]\d{0,3}x[1-9]\d{0,3}$'
            example: "800x600"
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Custom resolution image
          headers:
            Content-Type:
              schema:
                type: string
              example: "image/jpeg"
            Content-Length:
              schema:
                type: integer
              example: 131072
            Cache-Control:
              schema:
                type: string
              example: "public, max-age=31536000, immutable"
          content:
            image/*:
              schema:
                type: string
                format: binary
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/images/{id}/convert/{type}:
    get:
      tags:
        - Images
      summary:
        Download custom image format
      description: |
        Download a custom resolution version of the image.
        
        **Resolution Format:**
        - Must match pattern: WIDTHxHEIGHT (e.g., "800x600", "1200x900")
        - Both width and height must be positive integers
        - Maximum dimension: 10,000 pixels
        - Resolution must have been requested during upload
        
        **Processing:**
        - Smart-fit algorithm maintains aspect ratio
        - White background for letterboxing/pillarboxing
        - Quality optimized for resolution size
        
      operationId: downloadCustomResolution
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - $ref: '#/components/parameters/Format'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Custom format image
          headers:
            Content-Type:
              schema:
                type: string
              example: "image/jpeg"
            Content-Length:
              schema:
                type: integer
              example: 131072
            Cache-Control:
              schema:
                type: string
              example: "public, max-age=31536000, immutable"
          content:
            image/*:
              schema:
                type: string
                format: binary
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when authentication is enabled.
        
        Two types of keys are supported:
        - **Read-Write Keys**: Can upload images and access all read operations
        - **Read-Only Keys**: Can only access download and info operations
  parameters:
    ImageId:
      name: id
      in: path
      required: true
      description: Unique UUID identifier of the image
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: ETag for conditional requests (returns 304 if not modified)
      schema:
        type: string
      example: '"abc123def456"'
    Range:
      name: Range
      in: header
      required: false
      description: Byte range for partial content requests
      schema:
        type: string
      example: "bytes=0-1023"
    Resolution:
      name: resolution
      in: path
      required: true
      description: Image resolution/size
      schema:
        type: string
        pattern: '^(original|thumbnail|[1-9]\d{0,3}x[1-9]\d{0,3})$'
      example: "thumbnail"
    Format:
      name: Format
      in: path
      required: true
      description: Format type of the image
      schema:
        type: string
        enum: ["jpeg", "png", "gif", "webp"]
      example: "jpeg"
  schemas:
...
    PresignedURLResponse:
      type: object
      required:
        - url
        - expires_at
        - expires_in
      properties:
        url:
          type: string
          format: uri
          description: Presigned URL for direct access to the image
          example: "https://bucket.s3.amazonaws.com/images/f47ac10b-58cc-4372-a567-0e02b2c3d479/thumbnail.jpg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=..."
        expires_at:
          type: string
          format: date-time
          description: Timestamp when the presigned URL expires
          example: "2025-09-12T16:30:00Z"
        expires_in:
          type: integer
          description: URL expiration time in seconds
          example: 3600
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          description: Error category or type
          example: "Validation failed"
        message:
          type: string
          description: Human-readable error message
          example: "The uploaded file exceeds the maximum size limit"
        code:
          type: integer
          description: HTTP status code
          example: 400
  responses:
    BadRequest:
      description: Bad request - invalid input
      headers:
        X-Request-ID:
          schema:
            type: string
            format: uuid
          description: Unique request identifier for tracing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            InvalidFile:
              summary: Invalid file format
              value:
                error: "Validation failed"
                message: "Unsupported file format. Supported formats: JPEG, PNG, GIF, WebP"
                code: 400
            InvalidResolution:
              summary: Invalid resolution format
              value:
                error: "Validation failed"
                message: "Invalid resolution format. Use WIDTHxHEIGHT (e.g., 800x600)"
                code: 400
            InvalidUUID:
              summary: Invalid UUID
              value:
                error: "Validation failed"
                message: "Invalid UUID format"
                code: 400
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          schema:
            type: string
            format: uuid
          description: Unique request identifier for tracing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ImageNotFound:
              summary: Image not found
              value:
                error: "Resource not found"
                message: "Image with ID 'f47ac10b-58cc-4372-a567-0e02b2c3d479' not found"
                code: 404
            ResolutionNotFound:
              summary: Resolution not available
              value:
                error: "Resource not found"
                message: "Resolution '800x600' not available for this image"
                code: 404
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          schema:
            type: string
            format: uuid
          description: Unique request identifier for tracing
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
          example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Rate limit exceeded"
            message: "Too many requests. Please try again later"
            code: 429
    InternalServerError:
      description: Internal server error
      headers:
        X-Request-ID:
          schema:
            type: string
            format: uuid
          description: Unique request identifier for tracing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
            message: "An unexpected error occurred. Please try again later"
            code: 500
...
    NotModified:
      description: Not modified (304) - content unchanged
      headers:
        ETag:
          schema:
            type: string
          description: Entity tag
          example: '"abc123def456"'
        Cache-Control:
          schema:
            type: string
          description: Caching directives
          example: "public, max-age=31536000, immutable"
security: []
...